Don't change working directory to original dir when resuming run.
If we change the working directory to the one stored in the state file,
we'll continue to update the trees.zip at that location. This is
undesireable, as it means we can't copy the run directory to another
location and run from there (e.g., when testing a code change, or moving
to another filesystem location). Now, just trust that the user changed
to the desired working directory before invoking evolve.py.